Move bitset test utility
From Chaos helpers (excluded from build by default) to JetStream helpers
